router components refactor
