+++ src/app.js
@@ -1,66 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { BrowserRouter, Route, Switch, Link, NavLink } from "react-router-dom";
 import "normalize.css/normalize.css";
 import "./styles/styles.scss";
 
-const ExpenseDashboardPage = () => (
-  <div>
-    <p>Expense Dashboard</p>
-  </div>
-);
-const AddExpensePage = () => (
-  <div>
-    <p>Add Expense</p>
-  </div>
-);
-const EditExpensePage = () => (
-  <div>
-    <p>Edit Expense</p>
-  </div>
-);
-const HelpPage = () => (
-  <div>
-    <p>Help</p>
-  </div>
-);
-const ErrorPage = () => (
-  <div>
-    <p>
-      404! - <Link to="/">Dashboard</Link>
-    </p>
-  </div>
-);
-const Header = () => (
-  <header>
-    <h1>Expensify</h1>
-    <NavLink to="/" activeClassName="is-active" exact={true}>
-      Dashboard
-    </NavLink>
-    <NavLink to="/create" activeClassName="is-active">
-      Create
-    </NavLink>
-    <NavLink to="/edit" activeClassName="is-active">
-      Edit
-    </NavLink>
-    <NavLink to="/help" activeClassName="is-active">
-      Help
-    </NavLink>
-  </header>
-);
-const routes = (
-  <BrowserRouter>
-    <div>
-      <Header />
-      <Switch>
-        <Route path="/" component={ExpenseDashboardPage} exact={true} />
-        <Route path="/create" component={AddExpensePage} />
-        <Route path="/edit" component={EditExpensePage} />
-        <Route path="/help" component={HelpPage} />
-        <Route component={ErrorPage} />
-      </Switch>
-    </div>
-  </BrowserRouter>
-);
+import AppRouter from "./routers/AppRouter";
 
-ReactDOM.render(routes, document.getElementById("app"));
+ReactDOM.render(<AppRouter />, document.getElementById("app"));
