+++ src/components/AddExpensePage.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const AddExpensePage = () => (
+  <div>
+    <p>Add Expense</p>
+  </div>
+);
+
+export default AddExpensePage;
