+++ src/components/EditExpensePage.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const EditExpensePage = () => (
+  <div>
+    <p>Edit Expense</p>
+  </div>
+);
+
+export default EditExpensePage;
