+++ src/components/ExpenseDashboardPage.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const ExpenseDashboardPage = () => (
+  <div>
+    <p>Expense Dashboard</p>
+  </div>
+);
+
+export default ExpenseDashboardPage;
