+++ src/components/HelpPage.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const HelpPage = () => (
+  <div>
+    <p>Help</p>
+  </div>
+);
+
+export default HelpPage;
