+++ src/routers/AppRouter.js
@@ -1,0 +1,25 @@
+import React from "react";
+import { BrowserRouter, Route, Switch, Link, NavLink } from "react-router-dom";
+import Header from "../components/Header";
+import ErrorPage from "../components/ErrorPage";
+import HelpPage from "../components/HelpPage";
+import EditExpensePage from "../components/EditExpensePage";
+import AddExpensePage from "../components/AddExpensePage";
+import ExpenseDashboardPage from "../components/ExpenseDashboardPage";
+
+const AppRouter = () => (
+  <BrowserRouter>
+    <div>
+      <Header />
+      <Switch>
+        <Route path="/" component={ExpenseDashboardPage} exact={true} />
+        <Route path="/create" component={AddExpensePage} />
+        <Route path="/edit" component={EditExpensePage} />
+        <Route path="/help" component={HelpPage} />
+        <Route component={ErrorPage} />
+      </Switch>
+    </div>
+  </BrowserRouter>
+);
+
+export default AppRouter;
